# Revert fsnotify changes