--- a/hack/vendor.sh
+++ b/hack/vendor.sh
@@ -75,6 +75,9 @@ clone git github.com/fluent/fluent-logger-golang v1.0.0
 clone git github.com/philhofer/fwd 899e4efba8eaa1fea74175308f3fae18ff3319fa
 clone git github.com/tinylib/msgp 75ee40d2601edf122ef667e2a07d600d4c44490c
 
+# fsnotify
+clone git gopkg.in/fsnotify.v1 v1.2.0
+
 # awslogs deps
 clone git github.com/aws/aws-sdk-go v0.9.9
 clone git github.com/vaughan0/go-ini a98ad7ee00ec53921f08832bc06ecf7fd600e6a1