… vendor/src/gopkg.in/fsnotify.v1/.gitignore
… vendor/src/gopkg.in/fsnotify.v1/.travis.yml
… vendor/src/gopkg.in/fsnotify.v1/AUTHORS
… vendor/src/gopkg.in/fsnotify.v1/CHANGELOG.md
… vendor/src/gopkg.in/fsnotify.v1/CONTRIBUTING.md
… vendor/src/gopkg.in/fsnotify.v1/LICENSE
… vendor/src/gopkg.in/fsnotify.v1/README.md
… vendor/src/gopkg.in/fsnotify.v1/circle.yml
… vendor/src/gopkg.in/fsnotify.v1/fsnotify.go
… vendor/src/gopkg.in/fsnotify.v1/inotify.go
… vendor/src/gopkg.in/fsnotify.v1/inotify_poller.go
… vendor/src/gopkg.in/fsnotify.v1/kqueue.go
… vendor/src/gopkg.in/fsnotify.v1/open_mode_bsd.go
… vendor/src/gopkg.in/fsnotify.v1/open_mode_darwin.go
… vendor/src/gopkg.in/fsnotify.v1/windows.go