Reverts changes to remove fsnotify from json log reader.
This caused a slowdown that made TestLogsFollowSlowStdoutConsumer fail on ARM 100% of the time.

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkJSONFileLogger-8               53205         53077         -0.24%
BenchmarkJSONFileLoggerWithReader-8     93100         56483         -39.33%

benchmark                               old MB/s     new MB/s     speedup
BenchmarkJSONFileLogger-8               68.23        68.39        1.00x
BenchmarkJSONFileLoggerWithReader-8     20.30        33.46        1.65x

benchmark                               old allocs     new allocs     delta
BenchmarkJSONFileLogger-8               30             30             +0.00%
BenchmarkJSONFileLoggerWithReader-8     193            128            -33.68%

benchmark                               old bytes     new bytes     delta
BenchmarkJSONFileLogger-8               960           960           +0.00%
BenchmarkJSONFileLoggerWithReader-8     5753          3857          -32.96%
```