@mountkin 2016-02-24
@cpuguy83 can you check if #20605 mitigates the performance issue introduced by #19498 ?
In my test environment #20605 + #19498 is faster than the old approach.

fsnotify:
```
BenchmarkJSONFileLogger-8          	  200000	     67501 ns/op	  53.78 MB/s	     960 B/op	      30 allocs/op
BenchmarkJSONFileLoggerWithReader-8	  100000	    127740 ns/op	  14.80 MB/s	    5875 B/op	     197 allocs/op
```

#20605 + #19498
```
BenchmarkJSONFileLogger-8          	  300000	     57026 ns/op	  63.65 MB/s	     960 B/op	      30 allocs/op
BenchmarkJSONFileLoggerWithReader-8	  100000	    180476 ns/op	  10.47 MB/s	    2043 B/op	      80 allocs/op
```

BTW, I'm seeing the file poller is leaking a lot of fds on rotation (with the following test command)
```
docker run -d --name=test --log-opt max-size=50 --log-opt max-file=5 busybox sh -c "sleep 1;yes X|head -c 200"; docker logs -f test
```