@mountkin 2016-02-24
Can you help to test if https://github.com/mountkin/docker/tree/jsonlog-with-read-indexes fixed the ARM tests? I haven't got an ARM environment.