@mountkin 2016-02-24
I'll open another PR to see the ARM tests result :smiley: