@cpuguy83 2016-02-24
@mountkin Yeah, I see that we are missing a `defer f.Close()` in the pollers watch function... will follow-up after this PR if it gets merged.