@LK4D4 2016-02-24
Actually `docker logs -f` doesn't work with rotation at all in master("bad file descriptor").